+++ test/regressor/linear_regressor/linear_regressor_impl_test.dart
@@ -5,7 +5,6 @@ import 'package:ml_algo/src/regressor/linear_regressor/linear_regressor_constant
 import 'package:ml_algo/src/regressor/linear_regressor/linear_regressor_factory.dart';
 import 'package:ml_algo/src/regressor/linear_regressor/linear_regressor_impl.dart';
 import 'package:ml_dataframe/ml_dataframe.dart';
-import 'package:ml_linalg/dtype.dart';
 import 'package:ml_linalg/matrix.dart';
 import 'package:ml_linalg/vector.dart';
 import 'package:mockito/mockito.dart';